Fixed routing set-up
Fixed issues with loading pages from different routes by making the homepage route use '/recipes' as the beginning part of the route. Also needed to add --dirname + 'public' in app.js file so that the public folder was accessible to all routes. Every file linked in the public folder must be preceded with a '/'. For example '/images/banner.jpg' instead of images/banner.jpg